refactor(fpss): collapse the streaming wait strategy to the single low-latency behavior
The four-preset event-ring wait strategy (LowLatency / Balanced / Efficient / BusySpin) and its per-phase spin/yield/park tuning were an untuned knob: the shipped default was LowLatency everywhere and no caller had reason to leave it. Remove the whole tunability surface and hard-wire the fixed low-latency wait (100 spins, 10 yields, then a spin_loop hint; never sleeps), which is the correct trade-off for real-time market data.

`AdaptiveWaitStrategy` becomes a zero-sized type keeping only `low_latency()` / `wait_for()`; the connection path threads it unchanged. Dropped across the core: the `WaitMode` enum, the Balanced/Efficient/BusySpin presets, `with_tuning` / `from_mode`, the `MAX_SPIN_ITERS` / `MAX_YIELD_ITERS` / `MAX_PARK_US` caps, the `StreamingWaitStrategy` config enum with its Display/FromStr, the `wait_spin_iters` / `wait_yield_iters` / `wait_park_us` fields and their validation bounds, the `wait_strategy` / `wait_strategy_tuning` builder methods, and the `build_wait_strategy` bridge.

Removed the accessor surface from every binding so no hand-written projection outlives the core: the config_surface codegen entries plus regenerated FFI/py/ts/cpp accessors, the FFI `set_wait_spin_iters` / `set_wait_yield_iters` / `set_wait_park_us` entry points, the C header `THETADATADX_WAIT_*` constants and wait declarations, the C++ / py / ts wrappers and `.pyi` / `.d.ts` entries, and the parity rows. Wait-strategy round-trip tests are retargeted to the consumer-cpu coverage they shared. Reconnect `wait_*` cadences are untouched.

The Rust-only generic `StreamingClient::for_each_with_wait_strategy` escape hatch (any `W: WaitStrategy`) is kept.
