+++ docs-site/docs/migration/v12-to-v13.md
@@ -678,42 +678,28 @@ greeks.underlying_timestamp_ms
 
 Rust exposes the same names as methods returning `Option<i64>`; C exposes `thetadatadx_timestamp_ms(date, ms_of_day)` (returns `-1` for absent dates); C++ wraps it as `thetadatadx::timestamp_ms` returning `std::optional<int64_t>`. TypeScript precomputes the matching fields (`timestampMs`, `createdTimestampMs`, `lastTradeTimestampMs`, `underlyingTimestampMs`, `quoteTimestampMs`). List endpoints additionally return their values sorted ascending (numeric-aware for strike and date lists).
 
-## New: streaming latency tuning — wait strategy + consumer-core pinning
+## New: streaming consumer-core pinning
 
-This is an addition, not a breaking change: the defaults reproduce the previous behaviour exactly, so existing code needs no edits. Two opt-in knobs on the streaming config let you trade latency against CPU for the event-ring consumer.
+This is an addition, not a breaking change: the default reproduces the previous behaviour exactly, so existing code needs no edits. The event-ring consumer runs a fixed low-latency wait (spin, yield, then a `spin_loop` hint — never sleeps), and one opt-in knob lets you pin its thread to a CPU core.
 
-The wait strategy is the behaviour applied each time the consumer finds the ring momentarily empty:
-
-| Preset | Behaviour | Trade-off |
-|---|---|---|
-| `low_latency` (default) | spin, yield, `spin_loop` hint — never sleeps | lowest latency, highest idle CPU; identical to the previous fixed behaviour |
-| `balanced` | short spin then a brief park | low idle CPU, ~one park interval of added tail latency |
-| `efficient` | minimal spin then a longer park | lowest idle CPU |
-| `busy_spin` | pure spin, no yield or sleep | absolute minimum latency; pins a core while idle |
-
-The spin / yield / park counts tune independently (`wait_spin_iters`, `wait_yield_iters`, `wait_park_us`). The optional `consumer_cpu` knob pins the tick-consumer thread to a CPU core for deterministic, low-jitter delivery; leave it unset (the default) to keep the thread under the OS scheduler.
+The optional `consumer_cpu` knob pins the tick-consumer thread to a CPU core for deterministic, low-jitter delivery; leave it unset (the default) to keep the thread under the OS scheduler.
 
 ```python
 cfg = Config.production()
-cfg.wait_strategy = "balanced"   # "low_latency" | "balanced" | "efficient" | "busy_spin"
-cfg.wait_park_us = 100
 cfg.consumer_cpu = 3             # None (default) leaves it unpinned
 ```
 
 ```typescript
 const cfg = Config.production();
-cfg.setWaitStrategy("efficient");
 cfg.setConsumerCpu(3);          // null (default) leaves it unpinned
 ```
 
 ```cpp
 auto cfg = thetadatadx::Config::production();
-cfg.set_wait_strategy(THETADATADX_WAIT_BALANCED);   // 0..3, or use the constants
 cfg.set_consumer_cpu(3);                            // THETADATADX_CONSUMER_CPU_UNPINNED (-1) = unpinned
 ```
 
 ```c
-thetadatadx_config_set_wait_strategy(cfg, THETADATADX_WAIT_BALANCED);
 thetadatadx_config_set_consumer_cpu(cfg, 3);  /* negative = unpinned */
 ```
 
@@ -724,4 +710,4 @@ use thetadatadx::streaming::wait::BusySpin;
 client.for_each_with_wait_strategy(|event| handle(event), BusySpin);
 ```
 
-This generic override is Rust-only by design: `wait_for` fires on every ring-empty poll on the hot path, so routing a per-poll callback across the C ABI, the CPython lock, or the JavaScript event loop would be a latency regression. The preset enum plus the numeric tuning are the FFI-safe form for the other bindings.
+This generic override is Rust-only by design: `wait_for` fires on every ring-empty poll on the hot path, so routing a per-poll callback across the C ABI, the CPython lock, or the JavaScript event loop would be a latency regression. The bindings run the fixed low-latency wait.
