+++ thetadatadx-cpp/tests/error_taxonomy.cpp
@@ -212,10 +212,6 @@ TEST_CASE("config enum setters reject an out-of-domain value with InvalidParamet
     REQUIRE_THROWS_AS(cfg.set_flush_mode(9), thetadatadx::InvalidParameterError);
     REQUIRE_THROWS_AS(cfg.set_flush_mode(9), thetadatadx::ThetaDataError);
 
-    REQUIRE_NOTHROW(cfg.set_wait_strategy(THETADATADX_WAIT_BUSY_SPIN));
-    REQUIRE_THROWS_AS(cfg.set_wait_strategy(9), thetadatadx::InvalidParameterError);
-    REQUIRE_THROWS_AS(cfg.set_wait_strategy(9), thetadatadx::ThetaDataError);
-
     REQUIRE_NOTHROW(cfg.set_reconnect_jitter(2));
     REQUIRE_THROWS_AS(cfg.set_reconnect_jitter(9), thetadatadx::InvalidParameterError);
