+++ thetadatadx-cpp/tests/lifecycle.cpp
@@ -157,27 +157,11 @@ TEST_CASE("Config flush_mode getter round-trips", "[lifecycle][offline]") {
     REQUIRE(config.get_flush_mode() == 0);
 }
 
-TEST_CASE("Config wait_strategy / tuning / consumer_cpu round-trip", "[lifecycle][offline]") {
-    // Mirrors the Python `Config.wait_strategy` / `.consumer_cpu` and
-    // TypeScript `waitStrategy` / `consumerCpu` surfaces: a value set
-    // through the C++ wrapper reads back through the same wrapper.
+TEST_CASE("Config consumer_cpu round-trip", "[lifecycle][offline]") {
+    // Mirrors the Python `Config.consumer_cpu` and TypeScript
+    // `consumerCpu` surfaces: a value set through the C++ wrapper reads
+    // back through the same wrapper.
     auto config = thetadatadx::Config::production();
-
-    // Default preset is LowLatency (preserves the historical behaviour).
-    REQUIRE(config.get_wait_strategy() == THETADATADX_WAIT_LOW_LATENCY);
-
-    for (int mode : {THETADATADX_WAIT_LOW_LATENCY, THETADATADX_WAIT_BALANCED,
-                     THETADATADX_WAIT_EFFICIENT, THETADATADX_WAIT_BUSY_SPIN}) {
-        config.set_wait_strategy(mode);
-        REQUIRE(config.get_wait_strategy() == mode);
-    }
-
-    config.set_wait_spin_iters(16);
-    REQUIRE(config.get_wait_spin_iters() == 16);
-    config.set_wait_yield_iters(2);
-    REQUIRE(config.get_wait_yield_iters() == 2);
-    config.set_wait_park_us(200);
-    REQUIRE(config.get_wait_park_us() == 200);
 
     // Default consumer cpu is unpinned (the negative sentinel).
     REQUIRE(config.get_consumer_cpu() == THETADATADX_CONSUMER_CPU_UNPINNED);
