+++ thetadatadx-ffi/src/auth.rs
@@ -389,144 +389,6 @@ pub unsafe extern "C" fn thetadatadx_config_get_streaming_environment(
     })
 }
 
-/// Set the wait-strategy spin iteration count on a config handle.
-/// Returns `0` on success, `-1` (code `THETADATADX_ERR_CONFIG`) on a
-/// null handle.
-#[no_mangle]
-pub unsafe extern "C" fn thetadatadx_config_set_wait_spin_iters(
-    config: *mut ThetaDataDxConfig,
-    iters: u32,
-) -> i32 {
-    ffi_boundary!(-1, {
-        if config.is_null() {
-            crate::error::set_error_with_code(
-                "thetadatadx_config_set_wait_spin_iters: config handle is null",
-                crate::error::THETADATADX_ERR_CONFIG,
-            );
-            return -1;
-        }
-        // SAFETY: see `thetadatadx_config_set_flush_mode`.
-        let config = unsafe { &mut *config };
-        config.inner.streaming.wait_spin_iters = iters;
-        0
-    })
-}
-
-/// Read the wait-strategy spin iteration count. Returns `0` on success,
-/// `-1` if either pointer is null.
-#[no_mangle]
-pub unsafe extern "C" fn thetadatadx_config_get_wait_spin_iters(
-    config: *const ThetaDataDxConfig,
-    out_iters: *mut u32,
-) -> i32 {
-    ffi_boundary!(-1, {
-        if config.is_null() || out_iters.is_null() {
-            set_error("config or out-parameter pointer is null");
-            return -1;
-        }
-        // SAFETY: see `thetadatadx_config_get_flush_mode`.
-        let config = unsafe { &*config };
-        let value = config.inner.streaming.wait_spin_iters;
-        // SAFETY: out pointer checked non-null above.
-        unsafe {
-            *out_iters = value;
-        }
-        0
-    })
-}
-
-/// Set the wait-strategy yield iteration count on a config handle.
-/// Returns `0` on success, `-1` (code `THETADATADX_ERR_CONFIG`) on a
-/// null handle.
-#[no_mangle]
-pub unsafe extern "C" fn thetadatadx_config_set_wait_yield_iters(
-    config: *mut ThetaDataDxConfig,
-    iters: u32,
-) -> i32 {
-    ffi_boundary!(-1, {
-        if config.is_null() {
-            crate::error::set_error_with_code(
-                "thetadatadx_config_set_wait_yield_iters: config handle is null",
-                crate::error::THETADATADX_ERR_CONFIG,
-            );
-            return -1;
-        }
-        // SAFETY: see `thetadatadx_config_set_flush_mode`.
-        let config = unsafe { &mut *config };
-        config.inner.streaming.wait_yield_iters = iters;
-        0
-    })
-}
-
-/// Read the wait-strategy yield iteration count. Returns `0` on success,
-/// `-1` if either pointer is null.
-#[no_mangle]
-pub unsafe extern "C" fn thetadatadx_config_get_wait_yield_iters(
-    config: *const ThetaDataDxConfig,
-    out_iters: *mut u32,
-) -> i32 {
-    ffi_boundary!(-1, {
-        if config.is_null() || out_iters.is_null() {
-            set_error("config or out-parameter pointer is null");
-            return -1;
-        }
-        // SAFETY: see `thetadatadx_config_get_flush_mode`.
-        let config = unsafe { &*config };
-        let value = config.inner.streaming.wait_yield_iters;
-        // SAFETY: out pointer checked non-null above.
-        unsafe {
-            *out_iters = value;
-        }
-        0
-    })
-}
-
-/// Set the wait-strategy park interval (microseconds) on a config
-/// handle, used by the Balanced / Efficient strategies. Returns `0` on
-/// success, `-1` (code `THETADATADX_ERR_CONFIG`) on a null handle.
-#[no_mangle]
-pub unsafe extern "C" fn thetadatadx_config_set_wait_park_us(
-    config: *mut ThetaDataDxConfig,
-    park_us: u64,
-) -> i32 {
-    ffi_boundary!(-1, {
-        if config.is_null() {
-            crate::error::set_error_with_code(
-                "thetadatadx_config_set_wait_park_us: config handle is null",
-                crate::error::THETADATADX_ERR_CONFIG,
-            );
-            return -1;
-        }
-        // SAFETY: see `thetadatadx_config_set_flush_mode`.
-        let config = unsafe { &mut *config };
-        config.inner.streaming.wait_park_us = park_us;
-        0
-    })
-}
-
-/// Read the wait-strategy park interval in microseconds. Returns `0` on
-/// success, `-1` if either pointer is null.
-#[no_mangle]
-pub unsafe extern "C" fn thetadatadx_config_get_wait_park_us(
-    config: *const ThetaDataDxConfig,
-    out_park_us: *mut u64,
-) -> i32 {
-    ffi_boundary!(-1, {
-        if config.is_null() || out_park_us.is_null() {
-            set_error("config or out-parameter pointer is null");
-            return -1;
-        }
-        // SAFETY: see `thetadatadx_config_get_flush_mode`.
-        let config = unsafe { &*config };
-        let value = config.inner.streaming.wait_park_us;
-        // SAFETY: out pointer checked non-null above.
-        unsafe {
-            *out_park_us = value;
-        }
-        0
-    })
-}
-
 /// Pin the streaming consumer thread to a CPU core, or leave it under
 /// the OS scheduler.
 ///
