+++ thetadatadx-ffi/src/config_accessors.rs
@@ -1529,88 +1529,6 @@ pub unsafe extern "C" fn thetadatadx_config_get_flush_mode(
     })
 }
 
-/// Set the streaming event-ring consumer wait strategy on a config
-/// handle.
-///
-/// - `mode = 0`: LowLatency (default) — spin, yield, then a `spin_loop`
-///   hint; never sleeps. Lowest latency, highest idle CPU.
-/// - `mode = 1`: Balanced — short spin then a brief timed park. Low idle
-///   CPU, slightly higher tail latency.
-/// - `mode = 2`: Efficient — minimal spin then a longer timed park.
-///   Lowest idle CPU.
-/// - `mode = 3`: BusySpin — pure spin, no yield or sleep. Absolute
-///   minimum latency; pins a core while idle.
-///
-/// Returns `0` on success. Returns `-1` and sets `thetadatadx_last_error`
-/// when `mode` is outside the documented `{0, 1, 2, 3}` set (code
-/// `THETADATADX_ERR_INVALID_PARAMETER`) or when `config` is null (code
-/// `THETADATADX_ERR_CONFIG`).
-#[no_mangle]
-pub unsafe extern "C" fn thetadatadx_config_set_wait_strategy(
-    config: *mut ThetaDataDxConfig,
-    mode: i32,
-) -> i32 {
-    ffi_boundary!(-1, {
-        if config.is_null() {
-            crate::error::set_error_with_code(
-                "thetadatadx_config_set_wait_strategy: config handle is null",
-                crate::error::THETADATADX_ERR_CONFIG,
-            );
-            return -1;
-        }
-        let value = match mode {
-            0 => thetadatadx::StreamingWaitStrategy::LowLatency,
-            1 => thetadatadx::StreamingWaitStrategy::Balanced,
-            2 => thetadatadx::StreamingWaitStrategy::Efficient,
-            3 => thetadatadx::StreamingWaitStrategy::BusySpin,
-            other => {
-                crate::error::set_error_with_code(
-                    &format!(
-                        "thetadatadx_config_set_wait_strategy: invalid mode {other}; expected 0 (LowLatency), 1 (Balanced), 2 (Efficient), or 3 (BusySpin)"
-                    ),
-                    crate::error::THETADATADX_ERR_INVALID_PARAMETER,
-                );
-                return -1;
-            }
-        };
-        // SAFETY: config is a non-null pointer returned by `thetadatadx_config_*` and not yet freed; `&mut *` produces a unique reference valid for the call duration because the caller owns the Box and the FFI contract forbids concurrent calls on the same handle.
-        let config = unsafe { &mut *config };
-        config.inner.streaming.wait_strategy = value;
-        0
-    })
-}
-
-/// Read the configured streaming wait strategy. Same encoding as
-/// `thetadatadx_config_set_wait_strategy`: writes `0` (LowLatency), `1`
-/// (Balanced), `2` (Efficient), or `3` (BusySpin) into `*out_mode`.
-/// Returns `0` on success, `-1` if either pointer is null.
-#[no_mangle]
-pub unsafe extern "C" fn thetadatadx_config_get_wait_strategy(
-    config: *const ThetaDataDxConfig,
-    out_mode: *mut i32,
-) -> i32 {
-    ffi_boundary!(-1, {
-        if config.is_null() || out_mode.is_null() {
-            set_error("config or out-parameter pointer is null");
-            return -1;
-        }
-        // SAFETY: config is a non-null `*const ThetaDataDxConfig` returned by `thetadatadx_config_*` and not yet freed; `&*` produces a shared reference valid for the call duration.
-        let config = unsafe { &*config };
-        let value = match config.inner.streaming.wait_strategy {
-            thetadatadx::StreamingWaitStrategy::LowLatency => 0,
-            thetadatadx::StreamingWaitStrategy::Balanced => 1,
-            thetadatadx::StreamingWaitStrategy::Efficient => 2,
-            thetadatadx::StreamingWaitStrategy::BusySpin => 3,
-            _ => 0,
-        };
-        // SAFETY: out pointer checked non-null above; the FFI contract pins the storage for the call duration and forbids concurrent calls on the same handle.
-        unsafe {
-            *out_mode = value;
-        }
-        0
-    })
-}
-
 /// Set the jitter strategy applied to every reconnect delay.
 ///
 /// - `mode = 0`: Full (default) — sample uniformly from `[0, delay]`.
