+++ thetadatadx-rs/config_surface.toml
@@ -1970,8 +1970,8 @@ Current historical gRPC host.
 # variant { Variant => N }`); C++ passes the int through; Python / TS
 # speak the lowercase string via the core enum `parse` / `as_str`. A
 # rejected int carries `THETADATADX_ERR_INVALID_PARAMETER`. The
-# null-handle code is per-setter (`null_err`): flush_mode / wait_strategy
-# pin `THETADATADX_ERR_CONFIG`, jitter / host_selection fall back to the
+# null-handle code is per-setter (`null_err`): flush_mode pins
+# `THETADATADX_ERR_CONFIG`, jitter / host_selection fall back to the
 # untyped `THETADATADX_ERR_OTHER`.
 
 [[accessor]]
@@ -2071,128 +2071,6 @@ label = "batched"
 int = 1
 rust = "Immediate"
 label = "immediate"
-
-[[accessor]]
-symbol = "thetadatadx_config_set_wait_strategy"
-kind = "enum"
-param = "mode"
-abi_type = "i32"
-path = "streaming.wait_strategy"
-enum_type = "thetadatadx::StreamingWaitStrategy"
-enum_core = "config::StreamingWaitStrategy"
-enum_expected = "expected 0 (LowLatency), 1 (Balanced), 2 (Efficient), or 3 (BusySpin)"
-py_err = 'wait_strategy must be "low_latency", "balanced", "efficient", or "busy_spin"; got {strategy:?}'
-ts_err = 'setWaitStrategy: strategy must be "low_latency", "balanced", "efficient", or "busy_spin"; got {strategy:?}'
-doc = """
-Set the streaming event-ring consumer wait strategy on a config
-handle.
-
-- `mode = 0`: LowLatency (default) — spin, yield, then a `spin_loop`
-  hint; never sleeps. Lowest latency, highest idle CPU.
-- `mode = 1`: Balanced — short spin then a brief timed park. Low idle
-  CPU, slightly higher tail latency.
-- `mode = 2`: Efficient — minimal spin then a longer timed park.
-  Lowest idle CPU.
-- `mode = 3`: BusySpin — pure spin, no yield or sleep. Absolute
-  minimum latency; pins a core while idle.
-
-Returns `0` on success. Returns `-1` and sets `thetadatadx_last_error`
-when `mode` is outside the documented `{0, 1, 2, 3}` set (code
-`THETADATADX_ERR_INVALID_PARAMETER`) or when `config` is null (code
-`THETADATADX_ERR_CONFIG`).
-"""
-cpp_doc = """
-Set the streaming event-ring consumer wait strategy.
-0=LowLatency (default), 1=Balanced, 2=Efficient, 3=BusySpin
-(see the @c THETADATADX_WAIT_* constants). Throws
-@c thetadatadx::InvalidParameterError when @p mode is outside the
-documented `{0, 1, 2, 3}` set (and @c thetadatadx::ThetaDataError
-on a null handle).
-"""
-py_doc = """
-Set the streaming event-ring consumer wait strategy — the
-latency-vs-CPU knob applied on each ring-empty poll.
-
-Accepts ``"low_latency"`` (default, never sleeps — lowest
-latency, highest idle CPU), ``"balanced"`` (brief park — low idle
-CPU), ``"efficient"`` (longer park — lowest idle CPU), or
-``"busy_spin"`` (pure spin — pins a core). Tune the individual
-spin / yield / park counts via ``wait_spin_iters`` /
-``wait_yield_iters`` / ``wait_park_us``.
-"""
-ts_doc = """
-Set the streaming event-ring consumer wait strategy — the
-latency-vs-CPU knob applied on each ring-empty poll.
-
-Accepts `"low_latency"` (default — never sleeps, lowest latency,
-highest idle CPU), `"balanced"` (brief park — low idle CPU),
-`"efficient"` (longer park — lowest idle CPU), or `"busy_spin"`
-(pure spin — pins a core). Tune the spin / yield / park counts via
-`setWaitSpinIters` / `setWaitYieldIters` / `setWaitParkUs`.
-"""
-py_param = "strategy"
-ts_param = "strategy"
-[[accessor.variant]]
-int = 0
-rust = "LowLatency"
-label = "low_latency"
-[[accessor.variant]]
-int = 1
-rust = "Balanced"
-label = "balanced"
-[[accessor.variant]]
-int = 2
-rust = "Efficient"
-label = "efficient"
-[[accessor.variant]]
-int = 3
-rust = "BusySpin"
-label = "busy_spin"
-
-[[accessor]]
-symbol = "thetadatadx_config_get_wait_strategy"
-kind = "enum"
-param = "out_mode"
-abi_type = "i32"
-path = "streaming.wait_strategy"
-enum_type = "thetadatadx::StreamingWaitStrategy"
-enum_core = "config::StreamingWaitStrategy"
-# A future core variant decodes to `0` (LowLatency), the documented default.
-catch_all_int = 0
-doc = """
-Read the configured streaming wait strategy. Same encoding as
-`thetadatadx_config_set_wait_strategy`: writes `0` (LowLatency), `1`
-(Balanced), `2` (Efficient), or `3` (BusySpin) into `*out_mode`.
-Returns `0` on success, `-1` if either pointer is null.
-"""
-cpp_doc = """
-Read the current streaming wait strategy. Same encoding as
-@c set_wait_strategy. Returns `0` (LowLatency) on a null handle.
-"""
-py_doc = """
-Current streaming wait strategy (``"low_latency"``,
-``"balanced"``, ``"efficient"``, or ``"busy_spin"``).
-"""
-ts_doc = """
-Current streaming wait strategy (`"low_latency"`, `"balanced"`,
-`"efficient"`, or `"busy_spin"`).
-"""
-[[accessor.variant]]
-int = 0
-rust = "LowLatency"
-label = "low_latency"
-[[accessor.variant]]
-int = 1
-rust = "Balanced"
-label = "balanced"
-[[accessor.variant]]
-int = 2
-rust = "Efficient"
-label = "efficient"
-[[accessor.variant]]
-int = 3
-rust = "BusySpin"
-label = "busy_spin"
 
 [[accessor]]
 symbol = "thetadatadx_config_set_reconnect_jitter"
