+++ thetadatadx-rs/src/fpss/connection.rs
@@ -73,9 +73,8 @@ pub(crate) struct ConnectWithStreamArgs<'a> {
     pub host_shuffle_seed: u64,
     pub ring_size: usize,
     pub flush_mode: StreamingFlushMode,
-    /// Resolved event-ring consumer wait strategy (preset + tuning, or a
-    /// Rust-supplied raw strategy). Mirrors
-    /// [`crate::config::StreamingConfig::wait_strategy`].
+    /// Fixed low-latency event-ring consumer wait strategy
+    /// ([`super::ring::AdaptiveWaitStrategy`]).
     pub wait_strategy: super::ring::AdaptiveWaitStrategy,
     /// Optional CPU core to pin the event-ring consumer thread to;
     /// `None` leaves it under the OS scheduler. Mirrors
